scripts/requirements-analytics.txt
google-analytics-data>=0.18.0
google-auth>=2.0.0
